backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
pydantic==2.9.2
pydantic-settings==2.5.2

# Excel handling
openpyxl==3.1.5
pandas==2.2.3

# API integrations
serpapi==0.1.5
firecrawl-py==1.5.1
boto3==1.35.36
playwright==1.48.0

# Database
redis==5.2.0

# Utilities
python-dotenv==1.0.1
requests==2.32.3
aiofiles==24.1.0